% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finder.R
\name{mod_finder_ui}
\alias{mod_finder_ui}
\alias{mod_finder_server}
\title{mod_finder_ui and mod_finder_server}
\usage{
mod_finder_ui(id)

mod_finder_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
