% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_universe.R
\name{mod_universe_ui}
\alias{mod_universe_ui}
\alias{mod_universe_server}
\title{mod_universe_ui and mod_universe_server}
\usage{
mod_universe_ui(id)

mod_universe_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
